refactor: restructure onboarding flow into screen-based hierarchy
- Move onboarding coordinator and state to screens/onboarding/
- Organize onboarding screens into screens/ subfolder
- Move managers to managers/ subfolder
- Update package declarations throughout onboarding flow
