+++ app/src/main/java/com/bitchat/android/ui/screens/onboarding/OnboardingState.kt
@@ -1,4 +1,4 @@
-package com.bitchat.android.onboarding
+package com.bitchat.android.ui.screens.onboarding
 
 enum class OnboardingState {
     CHECKING,
