+++ app/src/main/java/com/bitchat/android/ui/screens/onboarding/managers/BatteryOptimizationManager.kt
@@ -1,4 +1,4 @@
-package com.bitchat.android.onboarding
+package com.bitchat.android.ui.screens.onboarding.managers
 
 import android.content.Context
 import android.content.Intent
