+++ app/src/main/java/com/bitchat/android/ui/screens/onboarding/managers/BluetoothStatusManager.kt
@@ -1,4 +1,4 @@
-package com.bitchat.android.onboarding
+package com.bitchat.android.ui.screens.onboarding.managers
 
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothManager
