+++ app/src/main/java/com/bitchat/android/ui/screens/onboarding/managers/LocationStatusManager.kt
@@ -1,4 +1,4 @@
-package com.bitchat.android.onboarding
+package com.bitchat.android.ui.screens.onboarding.managers
 
 import android.content.BroadcastReceiver
 import android.content.Context
