+++ app/src/main/java/com/bitchat/android/ui/screens/onboarding/screens/LocationCheckScreen.kt
@@ -1,4 +1,4 @@
-package com.bitchat.android.onboarding
+package com.bitchat.android.ui.screens.onboarding.screens
 
 import androidx.compose.animation.core.*
 import androidx.compose.foundation.layout.*
@@ -15,6 +15,7 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.bitchat.android.ui.screens.onboarding.managers.LocationStatus
 
 /**
  * Screen shown when checking location services status or requesting location services enable
